refactors the SearchResultList
